Add stats

+++ storage/classe.go
@@ -1,8 +1,8 @@
 package storage
 
 type Classe struct {
-	Name string
-	School string
-	NbStudent uint8
-	Students []Student
+	Name      string
+	School    string
+	NbStudent uint
+	Students  []Student
 }
